+++ packages/terra-clinical-label-value-view/src/terra-dev-site/doc/example/LabelValueViewTextTermAndDef.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import LabelValueView from 'terra-clinical-label-value-view';
+
+const LabelValueViewText = () => (
+  <div>
+    <style>
+      {`
+        #dl-wrapper > dd {
+          margin-bottom: 14px;
+        }
+      `}
+    </style>
+    <dl id="dl-wrapper">
+      <LabelValueView label="Collected By" textValue="Noah Brown" isChildOfDescriptionList />
+      <LabelValueView label="Date" textValue="12/12/12 00:00AM" isChildOfDescriptionList />
+      <LabelValueView label="Blood Pressure" textValue="110/60 mmHg" isChildOfDescriptionList />
+      <LabelValueView label="Temperature" textValue="97 F" isChildOfDescriptionList />
+      <LabelValueView label="Heart Rate" textValue="80 BPM" isChildOfDescriptionList />
+    </dl>
+  </div>
+);
+
+export default LabelValueViewText;