+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Created a new prop `isChildOfDescriptionList` on the LabelValueView component.
+  * Updated the LabelValueView to be:
+  *     - a React.Fragment with a term and a definition when the LabelValueView is already inside a description list (when isChildOfDescriptionList is sent with a "true" value)
+  *     - a description list when isChildOfDescriptionList is not sent (undefined) or sent as "false"
+  * Updated the component examples for new prop `isChildOfDescriptionList`.
 
 ## 3.28.0 - (April 21, 2023)